migrated Core project to SDK Project, extracted tests to a Core.Tests and migrated to NUnit 3
