+++ NEbml.sln
@@ -1,0 +1,54 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Src", "Src", "{D3EB2C4D-19E6-4394-9C30-0E2A8243BC03}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Src\Core\Core.csproj", "{06EB187F-17EA-487A-8EA0-F045A8216725}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.Tests", "Src\Core.Tests\Core.Tests.csproj", "{E0349793-FD46-49FB-B9B7-DEB33499CC21}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Debug|x64.Build.0 = Debug|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Debug|x86.Build.0 = Debug|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Release|Any CPU.Build.0 = Release|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Release|x64.ActiveCfg = Release|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Release|x64.Build.0 = Release|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Release|x86.ActiveCfg = Release|Any CPU
+		{06EB187F-17EA-487A-8EA0-F045A8216725}.Release|x86.Build.0 = Release|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Debug|x64.Build.0 = Debug|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Debug|x86.Build.0 = Debug|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Release|x64.ActiveCfg = Release|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Release|x64.Build.0 = Release|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Release|x86.ActiveCfg = Release|Any CPU
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{06EB187F-17EA-487A-8EA0-F045A8216725} = {D3EB2C4D-19E6-4394-9C30-0E2A8243BC03}
+		{E0349793-FD46-49FB-B9B7-DEB33499CC21} = {D3EB2C4D-19E6-4394-9C30-0E2A8243BC03}
+	EndGlobalSection
+EndGlobal
