--- Src/Core/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("NEbml Core")]
-[assembly: AssemblyDescription("NEbml read/write library")]
-[assembly: AssemblyCulture("")]
-
-[assembly: AssemblyVersion("0.1.4.0")]
-[assembly: AssemblyFileVersion("0.1.4.0")]
