+++ Src/MkvTitleEdit/Matroska/MatroskaDtd.cs
@@ -1,4 +1,4 @@
-﻿/* Copyright (c) 2011 Oleg Zee
+﻿/* Copyright (c) 2011-2020 Oleg Zee
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
