--- Src/NEbml.sln
@@ -1,43 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Supplementary", "Supplementary", "{E04FAEF2-DE81-416F-8E0E-8172178962E7}"
-	ProjectSection(SolutionItems) = preProject
-		..\.gitignore = ..\.gitignore
-		..\README.md = ..\README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MkvTitleEdit", "MkvTitleEdit\MkvTitleEdit.csproj", "{F2B2D694-3F87-4578-B36E-8FAB7D63806D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{38D51CF9-B33C-434F-9E5C-44B2F8C652F2}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{D06B002E-8DDD-4AF2-A9C9-05A8BE20759E}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\NuGet.Config = .nuget\NuGet.Config
-		.nuget\NuGet.exe = .nuget\NuGet.exe
-		.nuget\NuGet.targets = .nuget\NuGet.targets
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F2B2D694-3F87-4578-B36E-8FAB7D63806D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F2B2D694-3F87-4578-B36E-8FAB7D63806D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F2B2D694-3F87-4578-B36E-8FAB7D63806D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F2B2D694-3F87-4578-B36E-8FAB7D63806D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{38D51CF9-B33C-434F-9E5C-44B2F8C652F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{38D51CF9-B33C-434F-9E5C-44B2F8C652F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{38D51CF9-B33C-434F-9E5C-44B2F8C652F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{38D51CF9-B33C-434F-9E5C-44B2F8C652F2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(MercurialSourceControlSolutionProperties) = preSolution
-		SolutionIsControlled = True
-		SolutionBindings = <Solution Location In Database>
-	EndGlobalSection
-EndGlobal
